+++ Crypto/Utilities/Utilities.swift
@@ -1,0 +1,31 @@
+//
+//  Utylities.swift
+//  Crypto
+//
+//  Created by Paweł Brzozowski on 08/04/2022.
+//
+
+import Foundation
+import UIKit
+
+extension UIImageView {
+    func load(url: URL) {
+        DispatchQueue.global().async { [weak self] in
+            if let data = try? Data(contentsOf: url) {
+                if let image = UIImage(data: data) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            }
+        }
+    }
+}
+
+extension Double {
+    /// Rounds the double to decimal places value
+    func rounded(toPlaces places:Int) -> Double {
+        let divisor = pow(10.0, Double(places))
+        return (self * divisor).rounded() / divisor
+    }
+}
